Fix(declaration): le champ "année de déclaration" est demandé en tout début de simulation
