+++ packages/app/src/globals.d.ts
@@ -3,6 +3,7 @@ export type AppState = {
     formValidated: FormState;
     nomEntreprise: string;
     trancheEffectifs: TrancheEffectifs;
+    anneeDeclaration: number | undefined;
     debutPeriodeReference: string;
     finPeriodeReference: string;
   };
@@ -71,7 +72,6 @@ export type AppState = {
   informationsComplementaires: {
     formValidated: FormState;
     dateConsultationCSE: string;
-    anneeDeclaration: number | undefined;
     datePublication: string;
     lienPublication: string;
   };
@@ -228,6 +228,7 @@ export type ActionType =
 export type ActionInformationsSimulationData = {
   nomEntreprise: string;
   trancheEffectifs: TrancheEffectifs;
+  anneeDeclaration: number | undefined;
   debutPeriodeReference: string;
   finPeriodeReference: string;
 };
@@ -312,7 +313,6 @@ export type ActionInformationsDeclarantData = {
 
 export type ActionInformationsComplementairesData = {
   dateConsultationCSE: string;
-  anneeDeclaration: number | undefined;
   datePublication: string;
   lienPublication: string;
 };
